Move FileSystemExtensionContainer to IStorageContainer
