+++ Source/Testably.Abstractions.Testing/FileSystemMock.FileSystemInfoMock.cs
@@ -104,8 +104,8 @@ public sealed partial class FileSystemMock
 			=> FileSystem.Path.GetExtension(Location.FullPath);
 
 		/// <inheritdoc cref="IFileSystem.IFileSystemInfo.ExtensionContainer" />
-		public IFileSystem.IFileSystemExtensionContainer ExtensionContainer { get; }
-			= new FileSystemExtensionContainer();
+		public IFileSystem.IFileSystemExtensionContainer ExtensionContainer
+			=> Container.ExtensionContainer;
 
 		/// <inheritdoc cref="IFileSystem.IFileSystemInfo.FullName" />
 		public string FullName => Location.FullPath;
